backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.6.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.5
Django==6.0.3
django-filter==24.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
ruff==0.15.8
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
